requirements.txt
requests>=2.27.1
beautifulsoup4>=4.10.0
python-docx>=0.8.11
selenium>=4.1.0


webdriver-manager>=3.5.2
